Fixed pre-existing bugs in the ways error flows and weights and biases are updated during back propagation.
The original code was wrong to treat the error as a separate entity. The
error flows backwards and is transformed by the activation functions. I
discovered this issue when testing the network against the MNIST dataset
using the ReLU activation function.

Also from a pre-existing bug, the weights and biases should be updated
after the error has been calculated.

Supporting ReLU was the motivation behind changing the way weights and
biases are initialized. Ideally, the random numbers would be taken from
a normal distribution. I'm looking for ways to take care of this simply
and under friendly licensing terms.

I also fixed a few small compatibility bugs.